jwt 인증후 Member Entity 를 가져오는게 아닌 MemberDetailsDTO 로 변경... 그에 따른 컨트롤러단 전부 수정
ParentId Integer -> Long 변경

MemberRole 체크하는부분 쪽  Entity 내부 함수로 변경

+++ src/main/java/io/dkargo/bulletinboard/controller/CategoryController.java
@@ -26,7 +26,7 @@ public class CategoryController {
     @GetMapping("")
     @ResponseStatus(HttpStatus.OK)
     @Secured({"ROLE_USER", "ROLE_ADMIN"})
-    public List<ResFindCategoryDTO> findCategoryByParentId(@RequestParam(required = false) Integer parentId) {
+    public List<ResFindCategoryDTO> findCategoryByParentId(@RequestParam(required = false) Long parentId) {
         return categoryService.findCategoryByParentIdOrderByCategoryNameAsc(parentId);
     }
