+++ src/main/java/io/dkargo/bulletinboard/repository/CategoryRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 @Repository
 public interface CategoryRepository extends JpaRepository<Category,Long> {
-    List<Category> findCategoriesByParentIdOrderByCategoryNameAsc(Integer parentId);
+    List<Category> findCategoriesByParentIdOrderByCategoryNameAsc(Long parentId);
 
-    List<Category> findCategoriesByParentId(Integer parentId);
+    List<Category> findCategoriesByParentId(Long parentId);
 
 }
